+++ src/app/shell/personal-cabinet/provider/create-employee/create-employee.component.html
@@ -83,19 +83,25 @@
           <p class="step-name">{{ 'TITLES.ACCESS_MANAGMENT' | translate }}</p>
         </ng-template>
         <form class="step form flex flex-col justify-between items-stretch">
-          <label class="step-label"
-            >{{ 'FORMS.LABELS.SELECT_WORKSHOPS' | translate }}
-            <span class="step-required">*</span>
-          </label>
-          <app-entity-checkbox-dropdown
-            [entities]="truncatedItems$ | async"
-            [declination]="WorkshopDeclination"
-            [entityControl]="entityControl"
-            labelByDefault="FORMS.LABELS.CHECK_WORKSHOPS"
-            dropdownContainerClass="create-admin"
-            (entityCheck)="onWorkshopsSelect($event)"
-            (userInteraction)="onWorkshopsChange()">
-          </app-entity-checkbox-dropdown>
+          <ng-container *ngIf="isDeputy; then Deputy; else Employee"></ng-container>
+          <ng-template #Deputy>
+            <label class="step-label">{{ 'ALL_WORKSHOPS_ACCESS' | translate }}</label>
+          </ng-template>
+          <ng-template #Employee>
+            <label class="step-label"
+              >{{ 'FORMS.LABELS.SELECT_WORKSHOPS' | translate }}
+              <span class="step-required">*</span>
+            </label>
+            <app-entity-checkbox-dropdown
+              [entities]="truncatedItems$ | async"
+              [declination]="WorkshopDeclination"
+              [entityControl]="entityControl"
+              labelByDefault="FORMS.LABELS.CHECK_WORKSHOPS"
+              dropdownContainerClass="create-admin"
+              (entityCheck)="onWorkshopsSelect($event)"
+              (userInteraction)="onWorkshopsChange()">
+            </app-entity-checkbox-dropdown>
+          </ng-template>
         </form>
 
         <ng-container *ngTemplateOutlet="AgreementsTmpl"></ng-container>
